+++ app/helpers/admin/content_groups_helper.rb
@@ -1,0 +1,2 @@
+module Admin::ContentGroupsHelper
+end