+++ app/models/admin/content_page.rb
@@ -1,0 +1,17 @@
+class Admin::ContentPage < ActiveRecord::Base
+  # Default Scope
+  # Constants
+  # Attribute related macros
+  # Associations
+  belongs_to :content_group
+  mount_uploader :photo, PhotoUploader
+
+  # Validations
+  # Callback
+  # Other macros
+  # Concerns
+  include Deletable
+
+  # Named scopes
+  # Methods
+end