+++ db/migrate/20151002214153_create_content_groups.rb
@@ -1,0 +1,13 @@
+class CreateContentGroups < ActiveRecord::Migration
+  def change
+    create_table :content_groups do |t|
+      t.string :title
+      t.string :slug
+      t.integer :position
+      t.text :description
+      t.boolean :deleted, default: false, null: false
+
+      t.timestamps null: false
+    end
+  end
+end