+++ db/migrate/20151002214331_create_content_pages.rb
@@ -1,0 +1,15 @@
+class CreateContentPages < ActiveRecord::Migration
+  def change
+    create_table :content_pages do |t|
+      t.string :title
+      t.string :slug
+      t.text :content
+      t.integer :position
+      t.string :photo
+      t.integer :content_group_id
+      t.boolean :deleted, default: false, null: false
+
+      t.timestamps null: false
+    end
+  end
+end