+++ CHANGELOG.md
@@ -8,3 +8,4 @@
   - Ability to create partners for static page
   - Ability to create consent forms and privacy policy
   - Ability to create dashboard highlights
+  - Ability to create static learn pages